+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2023-8-25
 - Changed `/todays_scores`'s timezone to use PST/PDT instead of UTC when retrieving the current date
 - Fixed bug where games with TBD start time would display "scheduled" time
+- Fixed bug where previously suspended games would display original start time instead of resumed time
 ## 2023-8-24
 - Condensed scoreboard elements to better fit within the embed
 ## 2023-8-23